Data.Type.Predicate

Description

Combinators for working with type-level predicates, along with typeclasses for canonical proofs and deciding functions.

Predicates

type Predicate k = k ~> Type

A type-level predicate in Haskell. We say that the predicate P :: Predicate k is true/satisfied by input x :: k if there exists a value of type P @@ x, and that it false/disproved if such a value cannot exist. (Where @@ is Apply, the singleton library's type-level function application for mathcable functions)

See Provable and Decidable for more information on how to use, prove and decide these predicates.

The kind k ~> Type is the kind of "matchable" type-level functions in Haskell. They are type-level functions that are encoded as dummy type constructors ("defunctionalization symbols") that can be decidedly "matched" on for things like typeclass instances.

There are two ways to define your own predicates:

Using the predicate combinators and predicate transformers in this library and the singletons library, which let you construct pre-made predicates and sometimes create predicates from other predicates.

Manually creating a data type that acts as a matchable predicate.

For an example of the latter, we can create the "not p" predicate, which takes a predicate p as input and returns the negation of the predicate:

-- First, create the data type with the kind signature you want
data Not :: Predicate k -> Predicate k

-- Then, write the Apply instance, to specify the type of the
-- witnesses of that predicate
instance Apply (Not p) a = (p @@ a) -> Void

See the source of Data.Type.Predicate and Data.Type.Predicate.Logic for simple examples of hand-made predicates. For example, we have the always-true predicate Evident:

data Evident :: Predicate k
instance Apply Evident a = Sing a

And the "and" predicate combinator:

data (&&&) :: Predicate k -> Predicate k -> Predicate k
instance Apply (p &&& q) a = (p @@ a, q @@ a)

Typically it is recommended to create predicates from the supplied predicate combinators (TyPred can be used for any type constructor to turn it into a predicate, for instance) whenever possible.

newtype Wit p a

A Wit p a is a value of type p @@ a --- that is, it is a proof or witness that p is satisfied for a.

Constructors

Wit

Fields

getWit :: p @@ a

Construct Predicates

type TyPred = (TyCon1 :: (k -> Type) -> Predicate k)

Convert a normal '->' type constructor into a Predicate.

TyPred :: (k -> Type) -> Predicate k

type Evident = (TyPred Sing :: Predicate k)

The always-true predicate.

Evident :: Predicate k

type EqualTo (a :: k) = (TyPred ((:~:) a) :: Predicate k)

EqualTo a is a predicate that the input is equal to a.

type BoolPred (p :: k ~> Bool) = (EqualTo True .@#@$$$ p :: Predicate k)

Convert a tradtional k ~> Bool predicate into a Predicate.

BoolPred :: (k ~> Bool) -> Predicate k

type Impossible = (Not Evident :: Predicate k)

The always-false predicate

Could also be defined as ConstSym1 Void, but this defintion gives us a free Decidable instance.

Manipulate predicates

type PMap (f :: k ~> j) (p :: Predicate j) = (p .@#@$$$ f :: Predicate k)

Pre-compose a function to a predicate

PMap :: (k ~> j) -> Predicate j -> Predicate k

data Not :: Predicate k -> Predicate k

Not p is the predicate that p is not true.

decideNot :: forall p a. Decision (p @@ a) -> Decision (Not p @@ a)

Decide Not p based on decisions of p.

Provable Predicates

type Prove p = forall a. Sing a -> p @@ a

A proving function for predicate p. See Provable for more information.

type (-->) p q = forall a. Sing a -> (p @@ a) -> q @@ a infixr 1

We say that p implies q (p --> q) if, given p a, we can always prove q @@ a.

type (-->#) p q h = forall a. Sing a -> (p @@ a) -> h (q @@ a) infixr 1

This is implication -->#, but only in a specific context h.

class Provable p where

A typeclass for provable predicates (constructivist tautologies).

A predicate is provable if, given any input a, you can generate a proof of p @@ a. Essentially, it means that a predicate is "always true".

This typeclass associates a canonical proof function for every provable predicate.

It confers two main advatnages:

The proof function for every predicate is available via the same name

We can write Provable instances for polymorphic predicate transformers (predicates parameterized on other predicates) easily, by refering to Provable instances of the transformed predicates.

Methods

prove :: Prove p

The canonical proving function for predicate p.

Note that prove is ambiguously typed, so you always need to call by specifying the predicate you want to prove using TypeApplications syntax:

prove @MyPredicate

type Disprovable p = Provable (Not p)

Disprovable p is a constraint that p can be disproven.

disprove :: forall p. Disprovable p => Prove (Not p)

The deciding/disproving function for Disprovable p.

Must be called by applying the Predicate to disprove:

disprove @p

type ProvableTC p = Provable (TyPred p)

If T :: k -> Type is a type constructor, then ProvableTC T is a constraint that T is "decidable", in that you have a canonical function:

proveTC :: Sing a -> T a

Is essentially Provable, except with type constructors k -> Type instead of matchable type-level functions (that are k ~> Type).

Since: 0.1.1.0

proveTC :: forall t a. ProvableTC t => Sing a -> t a

The canonical proving function for DecidableTC t.

Since: 0.1.1.0

class TFunctor f where

Implicatons p --> q can be lifted "through" a TFunctor into an f p --> f q.

Methods

tmap :: forall p q. (p --> q) -> f p --> f q

compImpl :: forall p q r. (p --> q) -> (q --> r) -> p --> r

Compose two implications.

Decidable Predicates

type Decide p = forall a. Sing a -> Decision (p @@ a)

A decision function for predicate p. See Decidable for more information.

type (-?>) p q = forall a. Sing a -> (p @@ a) -> Decision (q @@ a) infixr 1

Like implication -->, but knowing p @@ a can only let us decidably prove q a is true or false.

type (-?>#) p q h = forall a. Sing a -> (p @@ a) -> h (Decision (q @@ a)) infixr 1

Like -?>, but only in a specific context h.

class Decidable p where

A typeclass for decidable predicates.

A predicate is decidable if, given any input a, you can either prove or disprove p @@ a. A Decision (p @@ a) is a data type that has a branch p @@ a and Refuted (p @@ a).

This typeclass associates a canonical decision function for every decidable predicate.

It confers two main advatnages:

The decision function for every predicate is available via the same name

We can write Decidable instances for polymorphic predicate transformers (predicates parameterized on other predicates) easily, by refering to Decidable instances of the transformed predicates.

Minimal complete definition

Nothing

Methods

decide :: Decide p

The canonical decision function for predicate p.

Note that decide is ambiguously typed, so you always need to call by specifying the predicate you want to prove using TypeApplications syntax:

decide @MyPredicate

decide :: Provable p => Decide p

The canonical decision function for predicate p.

Note that decide is ambiguously typed, so you always need to call by specifying the predicate you want to prove using TypeApplications syntax:

decide @MyPredicate

type DecidableTC p = Decidable (TyPred p)

If T :: k -> Type is a type constructor, then DecidableTC T is a constraint that T is "decidable", in that you have a canonical function:

decideTC :: Sing a -> Decision (T a)

Is essentially Decidable, except with type constructors k -> Type instead of matchable type-level functions (that are k ~> Type).

Since: 0.1.1.0

decideTC :: forall t a. DecidableTC t => Sing a -> Decision (t a)

The canonical deciding function for DecidableTC t.

Since: 0.1.1.0

class DFunctor f where

Implicatons p -?> q can be lifted "through" a DFunctor into an f p -?> f q.

Methods

dmap :: forall p q. (p -?> q) -> f p -?> f q

Manipulate Decisions

data Decision a

A Decision about a type a is either a proof of existence or a proof that a cannot exist.

Constructors

Proved a
Witness for a

Disproved (Refuted a)
Proof that no a exists

flipDecision :: Decision a -> Decision (Refuted a)

Flip the contents of a decision. Turn a proof of a into a disproof of not-a.

Note that this is not reversible in general in Haskell. See doubleNegation for a situation where it is.

Since: 0.1.1.0

mapDecision :: (a -> b) -> (b -> a) -> Decision a -> Decision b

Map over the value inside a Decision.

forgetDisproof :: Decision a -> Maybe a

Converts a Decision to a Maybe. Drop the witness of disproof of a, returning Just if Proved (with the proof) and Nothing if Disproved.

Since: 0.1.1.0

forgetProof :: Decision a -> Maybe (Refuted a)

Drop the witness of proof of a, returning Nothing if Proved and Just if Disproved (with the disproof).

Since: 0.1.1.0

isProved :: Decision a -> Bool

Boolean test if a Decision is Proved.

Since: 0.1.1.0

isDisproved :: Decision a -> Bool

Boolean test if a Decision is Disproved.

Since: 0.1.1.0