Data.Type.Predicate.Auto

Description

Useful utilities for situations where you know that a predicate P is satisfied for a specific a at compile-time.

Since: 0.1.1.0

Automatically generate witnesses at compile-time

class Auto (p :: Predicate k) (a :: k) where

Automatically generate a witness for predicate p applied to input a.

Mostly useful for situations where you know a at compile-time, so you can just write auto directly in your source code. The choice is intended to mirror the auto keyword in languages like Idris.

Very close in nature to the Known typeclass in the type-combinators library.

Admittedly this interface is a bit clunky and ad-hoc; at this point you can just try writing auto in your code and praying that it works. You always have the option, of course, to just manually write proofs. If you have any inference rules to suggest, feel free to submit a PR!

An important limitation of Auto is the Haskell type system prevents "either-or" constraints; this could potentially be implemented using compiler plugins.

One consequence of this is that it is impossible to automatically derive Any f p and Not (All f p).

For these, the compiler needs help; you can use autoAny and autoNotAll for these situations.

Methods

auto :: p @@ a

Have the compiler generate a witness for p @@ a.

Must be called using type application syntax:

auto _ p @a

type AutoNot (p :: Predicate k) = Auto (Not p)

An AutoNot p a constraint means that p @@ a can be proven to not be true at compiletime.

Since: 0.1.2.0

autoNot :: forall k (p :: Predicate k) (a :: k). AutoNot p a => Not p @@ a

Disprove p @@ a at compiletime.

autoNot _ p a :: Not p @ a

Since: 0.1.2.0

data AutoProvable :: Predicate k -> Predicate k

Helper "predicate transformer" that gives you an instant auto for any Provable instance.

For example, say you have predicate P that you know is Provable, and you wish to generate a P @@ x, for some specific x you know at compile-time. You can use:

auto _ (AutoProvable P) @x

to obtain a P @@ x.

AutoProvable is essentially the identity function.

Helper classes

class AutoElem f (as :: f k) (a :: k) where

Typeclass representing Elems pointing to an a :: k that can be generated automatically from type-level collection as :: f k.

If GHC knows both the type-level collection and the element you want to find at compile-time, this instance should allow it to find it.

Used to help in the instance of Auto for the In predicate.

Example usage:

autoElem :: Index '[1,6,2,3] 2
-- IS (IS IZ)        -- third spot

And when used with Auto:

auto _ (In [] '[1,6,2,3]) @2
-- IS (IS IZ)

Methods

autoElem :: Elem f as a

Generate the Elem pointing to the a :: in a type-level collection as :: f k.

class AutoAll f (p :: Predicate k) (as :: f k) where

Helper class for deriving Auto instances for All predicates; each Universe instance is expected to implement these if possible, to get free Auto instaces for their All predicates.

Also helps for Not Any predicates and Not Found AnyMatch predicates.

Since: 0.1.2.0

Methods

autoAll :: All f p @@ as

Generate an All for a given predicate over all items in as.

Auto with help

autoAny :: forall f p as a. Auto p a => Elem f as a -> Any f p @@ as

Helper function to generate an Any f p if you can pick out a specific a in as where the predicate is provable at compile-time.

This is used to get around a fundamental limitation of Auto as a Haskell typeclass.

Since: 0.1.2.0

autoNotAll :: forall p f as a. (AutoNot p a, SingI as) => Elem f as a -> Not (All f p) @@ as

Helper function to generate a Not (All f p) if you can pick out a specific a in as where the predicate is disprovable at compile-time.

This is used to get around a fundamental limitation of Auto as a Haskell typeclass.

Since: 0.1.2.0