Data.Type.Universe

Description

A type family for "containers", intended for allowing lifting of predicates on k to be predicates on containers f k.

Universe

type family Elem (f :: Type -> Type) = (i :: f k -> k -> Type) | i -> f

type In (f :: Type -> Type) (as :: f k) = ElemSym1 f as

In f as is a predicate that a given input a is a member of collection as.

class FProd f => Universe (f :: Type -> Type) where

Typeclass for a type-level container that you can quantify or lift type-level predicates over.

Methods

idecideAny

Arguments

:: forall (p :: k ~> Type) (as :: f k). (forall a. Elem f as a -> Sing a -> Decision (p @@ a))
predicate on value

-> Sing as -> Decision (Any f p @@ as)
predicate on collection

decideAny, but providing an Elem.

idecideAll

Arguments

:: forall (p :: k ~> Type) (as :: f k). (forall a. Elem f as a -> Sing a -> Decision (p @@ a))
predicate on value

-> Sing as -> Decision (All f p @@ as)
predicate on collection

decideAll, but providing an Elem.

allProd :: forall p g. (forall a. Sing a -> (p @@ a) -> g a) -> All f p --> TyPred (Prod f g)

prodAll :: forall p g as. (forall a. g a -> p @@ a) -> Prod f g as -> All f p @@ as

singAll :: forall f k (as :: f k). Universe f => Sing as -> All f Evident @@ as

Split a Sing as into a proof that all a in as exist.

Instances

data Index (a :: [k]) (b :: k) :: forall k. [k] -> k -> Type where

Witness an item in a type-level list by providing its index.

The number of ISs correspond to the item's position in the list.

IZ         :: Index '[5,10,2] 5
IS IZ      :: Index '[5,10,2] 10
IS (IS IZ) :: Index '[5,10,2] 2

Constructors

IZ :: forall k (a :: [k]) (b :: k) (as :: [k]). Index (b ': as) b

IS :: forall k (a :: [k]) (b :: k) (bs :: [k]) (b1 :: k). Index bs b -> Index (b1 ': bs) b

data IJust (a :: Maybe k) (b :: k) :: forall k. Maybe k -> k -> Type where

Witness an item in a type-level Maybe by proving the Maybe is Just.

Constructors

IJust :: forall k (a :: Maybe k) (b :: k). IJust (Just b) b

data IRight (a :: Either j k) (b :: k) :: forall j k. Either j k -> k -> Type where

Witness an item in a type-level Either j by proving the Either is Right.

Constructors

IRight :: forall j k (a :: Either j k) (b :: k). IRight (Right b :: Either j k) b

data NEIndex (a :: NonEmpty k) (b :: k) :: forall k. NonEmpty k -> k -> Type where

Witness an item in a type-level NonEmpty by either indicating that it is the "head", or by providing an index in the "tail".

Constructors

NEHead :: forall k (a :: NonEmpty k) (b :: k) (as :: [k]). NEIndex (b :| as) b

NETail :: forall k (a :: NonEmpty k) (b :: k) (as :: [k]) (b1 :: k). Index as b -> NEIndex (b1 :| as) b

data ISnd (a :: (j, k)) (b :: k) :: forall j k. (j, k) -> k -> Type where

Trivially witness an item in the second field of a type-level tuple.

Constructors

ISnd :: forall j k (a :: (j, k)) (b :: k) (a1 :: j). ISnd ((,) a1 b) b

data IIdentity (a :: Identity k) (b :: k) :: forall k. Identity k -> k -> Type where

Trivially witness the item held in an Identity.

Since: functor-products-0.1.3.0

Constructors

IId :: forall k (a :: Identity k) (b :: k). IIdentity (Identity b) b

Predicates

data All f :: Predicate k -> Predicate (f k)

An All f p is a predicate testing a collection as :: f a for the fact that all items in as satisfy p. Represents the "forall" quantifier over a given universe.

This is mostly useful for its Decidable, Provable, and TFunctor instances, which lets you lift predicates on p to predicates on All f p.

newtype WitAll f p (as :: f k)

A WitAll p as is a witness that the predicate p a is true for all items a in the type-level collection as.

Constructors

WitAll

Fields

runWitAll :: forall a. Elem f as a -> p @@ a

type NotAll f p = (Not (All f p) :: Predicate (f k))

A NotAll f p is a predicate on a collection as that at least one a in as does not satisfy predicate p.

data Any f :: Predicate k -> Predicate (f k)

An Any f p is a predicate testing a collection as :: f a for the fact that at least one item in as satisfies p. Represents the "exists" quantifier over a given universe.

This is mostly useful for its Decidable and TFunctor instances, which lets you lift predicates on p to predicates on Any f p.

data WitAny f :: (k ~> Type) -> f k -> Type where

A WitAny p as is a witness that, for at least one item a in the type-level collection as, the predicate p a is true.

Constructors

WitAny :: Elem f as a -> (p @@ a) -> WitAny f p as

type None f p = (Not (Any f p) :: Predicate (f k))

A None f p is a predicate on a collection as that no a in as satisfies predicate p.

type Null f = (None f Evident :: Predicate (f k))

Predicate that a given as :: f k is empty and has no items in it.

type NotNull f = (Any f Evident :: Predicate (f k))

Predicate that a given as :: f k is not empty, and has at least one item in it.

Specialized

type IsJust = (NotNull Maybe :: Predicate (Maybe k))

Test that a Maybe is Just.

Since: 0.1.2.0

type IsNothing = (Null Maybe :: Predicate (Maybe k))

Test that a Maybe is Nothing.

Since: 0.1.2.0

type IsRight = (NotNull (Either j) :: Predicate (Either j k))

Test that an Either is Right

Since: 0.1.2.0

type IsLeft = (Null (Either j) :: Predicate (Either j k))

Test that an Either is Left

Since: 0.1.2.0

Decisions and manipulations

decideAny

Arguments

:: forall (p :: k ~> Type). Universe f

=> Decide p
predicate on value

-> Decide (Any f p)
predicate on collection

Lifts a predicate p on an individual a into a predicate that on a collection as that is true if and only if any item in as satisfies the original predicate.

That is, it turns a predicate of kind k ~> Type into a predicate of kind f k ~> Type.

Essentially tests existential quantification.

decideAll

Arguments

:: forall (p :: k ~> Type). Universe f

=> Decide p
predicate on value

-> Decide (All f p)
predicate on collection

Lifts a predicate p on an individual a into a predicate that on a collection as that is true if and only if all items in as satisfies the original predicate.

That is, it turns a predicate of kind k ~> Type into a predicate of kind f k ~> Type.

Essentially tests universal quantification.

genAll

Arguments

:: forall (p :: k ~> Type). Universe f

=> Prove p
always-true predicate on value

-> Prove (All f p)
always-true predicate on collection

If p a is true for all values a in as, then we have All p as. Basically witnesses the definition of All.

igenAll

Arguments

:: forall (p :: k ~> Type) (as :: f k). Universe f

=> (forall a. Elem f as a -> Sing a -> p @@ a)
always-true predicate on value

-> Sing as -> All f p @@ as
always-true predicate on collection

genAll, but providing an Elem.

splitSing :: forall f k (as :: f k). Universe f => Sing as -> All f (TyPred Sing) @@ as

Split a Sing as into a proof that all a in as exist.

pickElem :: forall f k (as :: f k) a. (Universe f, SingI as, SingI a, SDecide k) => Decision (Elem f as a)

Automatically generate a witness for a member, if possible