Data.Type.Universe

Description

A type family for "containers", intended for allowing lifting of predicates on k to be predicates on containers f k.

Universe

type family Elem (f :: Type -> Type) :: f k -> k -> Type

A witness for membership of a given item in a type-level collection

type In (f :: Type -> Type) (as :: f k) = ElemSym1 f as

In f as is a predicate that a given input a is a member of collection as.

class Universe (f :: Type -> Type) where

Typeclass for a type-level container that you can quantify or lift type-level predicates over.

Methods

idecideAny

Arguments

:: forall (p :: k ~> Type) (as :: f k). (forall a. Elem f as a -> Sing a -> Decision (p @@ a))
predicate on value

-> Sing as -> Decision (Any f p @@ as)
predicate on collection

decideAny, but providing an Elem.

idecideAll

Arguments

:: forall (p :: k ~> Type) (as :: f k). (forall a. Elem f as a -> Sing a -> Decision (p @@ a))
predicate on value

-> Sing as -> Decision (All f p @@ as)
predicate on collection

decideAll, but providing an Elem.

igenAllA

Arguments

:: forall (p :: k ~> Type) (as :: f k). Applicative h

=> (forall a. Elem f as a -> Sing a -> h (p @@ a))
predicate on value in context

-> Sing as -> h (All f p @@ as)
predicate on collection in context

genAllA, but providing an Elem.

Instances

data Index :: [k] -> k -> Type where

Witness an item in a type-level list by providing its index.

Constructors

IZ :: Index (a ': as) a

IS :: Index bs a -> Index (b ': bs) a

data IJust :: Maybe k -> k -> Type where

Witness an item in a type-level Maybe by proving the Maybe is Just.

Constructors

IJust :: IJust (Just a) a

data IRight :: Either j k -> k -> Type where

Witness an item in a type-level Either j by proving the Either is Right.

Constructors

IRight :: IRight (Right a) a

data NEIndex :: NonEmpty k -> k -> Type where

Witness an item in a type-level NonEmpty by either indicating that it is the "head", or by providing an index in the "tail".

Constructors

NEHead :: NEIndex (a :| as) a

NETail :: Index as a -> NEIndex (b :| as) a

data ISnd :: (j, k) -> k -> Type where

Trivially witness an item in the second field of a type-level tuple.

Constructors

ISnd :: ISnd '(a, b) b

data IProxy :: Proxy k -> k -> Type

There are no items of type a in a Proxy a.

Since: 0.1.3.0

data IIdentity :: Identity k -> k -> Type where

Trivially witness the item held in an Identity.

Since: 0.1.3.0

Constructors

IId :: IIdentity (Identity x) x

data CompElem :: (f :.: g) k -> k -> Type where

A pair of indices allows you to index into a nested structure.

Since: 0.1.2.0

Constructors

(:?) :: Elem f ass as -> Elem g as a -> CompElem (Comp ass) a

data SumElem :: (f :+: g) k -> k -> Type where

Index into a disjoint union by providing an index into one of the two possible options.

Since: 0.1.3.0

Constructors

IInL :: Elem f as a -> SumElem (InL as) a

IInR :: Elem f bs b -> SumElem (InR bs) b

Predicates

data All f :: Predicate k -> Predicate (f k)

An All f p is a predicate testing a collection as :: f a for the fact that all items in as satisfy p. Represents the "forall" quantifier over a given universe.

This is mostly useful for its Decidable, Provable, and TFunctor instances, which lets you lift predicates on p to predicates on All f p.

newtype WitAll f p (as :: f k)

A WitAll p as is a witness that the predicate p a is true for all items a in the type-level collection as.

Constructors

WitAll

Fields

runWitAll :: forall a. Elem f as a -> p @@ a

type NotAll f p = (Not (All f p) :: Predicate (f k))

A NotAll f p is a predicate on a collection as that at least one a in as does not satisfy predicate p.

data Any f :: Predicate k -> Predicate (f k)

An Any f p is a predicate testing a collection as :: f a for the fact that at least one item in as satisfies p. Represents the "exists" quantifier over a given universe.

This is mostly useful for its Decidable and TFunctor instances, which lets you lift predicates on p to predicates on Any f p.

data WitAny f :: (k ~> Type) -> f k -> Type where

A WitAny p as is a witness that, for at least one item a in the type-level collection as, the predicate p a is true.

Constructors

WitAny :: Elem f as a -> (p @@ a) -> WitAny f p as

type None f p = (Not (Any f p) :: Predicate (f k))

A None f p is a predicate on a collection as that no a in as satisfies predicate p.

type Null f = (None f Evident :: Predicate (f k))

Predicate that a given as :: f k is empty and has no items in it.

type NotNull f = (Any f Evident :: Predicate (f k))

Predicate that a given as :: f k is not empty, and has at least one item in it.

Specialized

type IsJust = (NotNull Maybe :: Predicate (Maybe k))

Test that a Maybe is Just.

Since: 0.1.2.0

type IsNothing = (Null Maybe :: Predicate (Maybe k))

Test that a Maybe is Nothing.

Since: 0.1.2.0

type IsRight = (NotNull (Either j) :: Predicate (Either j k))

Test that an Either is Right

Since: 0.1.2.0

type IsLeft = (Null (Either j) :: Predicate (Either j k))

Test that an Either is Left

Since: 0.1.2.0

Decisions and manipulations

decideAny

Arguments

:: forall (p :: k ~> Type). Universe f

=> Decide p
predicate on value

-> Decide (Any f p)
predicate on collection

Lifts a predicate p on an individual a into a predicate that on a collection as that is true if and only if any item in as satisfies the original predicate.

That is, it turns a predicate of kind k ~> Type into a predicate of kind f k ~> Type.

Essentially tests existential quantification.

decideAll

Arguments

:: forall (p :: k ~> Type). Universe f

=> Decide p
predicate on value

-> Decide (All f p)
predicate on collection

Lifts a predicate p on an individual a into a predicate that on a collection as that is true if and only if all items in as satisfies the original predicate.

That is, it turns a predicate of kind k ~> Type into a predicate of kind f k ~> Type.

Essentially tests universal quantification.

genAllA

Arguments

:: forall (p :: k ~> Type) (as :: f k). (Universe f, Applicative h)

=> (forall a. Sing a -> h (p @@ a))
predicate on value in context

-> Sing as -> h (All f p @@ as)
predicate on collection in context

If p a is true for all values a in as under some (Applicative) context h, then you can create an All p as under that Applicative context h.

Can be useful with Identity (which is basically unwrapping and wrapping All), or with Maybe (which can express predicates that are either provably true or not provably false).

In practice, this can be used to iterate and traverse and sequence actions over all "items" in as.

genAll

Arguments

:: forall (p :: k ~> Type). Universe f

=> Prove p
always-true predicate on value

-> Prove (All f p)
always-true predicate on collection

If p a is true for all values a in as, then we have All p as. Basically witnesses the definition of All.

igenAll

Arguments

:: forall (p :: k ~> Type) (as :: f k). Universe f

=> (forall a. Elem f as a -> Sing a -> p @@ a)
always-true predicate on value

-> Sing as -> All f p @@ as
always-true predicate on collection

genAll, but providing an Elem.

foldMapUni :: forall f k (as :: f k) m. (Universe f, Monoid m) => (forall (a :: k). Sing a -> m) -> Sing as -> m

A foldMap over all items in a collection.

ifoldMapUni :: forall f k (as :: f k) m. (Universe f, Monoid m) => (forall a. Elem f as a -> Sing a -> m) -> Sing as -> m

foldMapUni but with access to the index.

index

Arguments

:: Universe f

=> Elem f as a
Witness

-> Sing as
Collection

-> Sing a

Extract the item from the container witnessed by the Elem

pickElem :: forall f k (as :: f k) a. (Universe f, SingI as, SingI a, SDecide k) => Decision (Elem f as a)

Automatically generate a witness for a member, if possible

Universe Combination

data family Sing (a :: k) :: Type

The singleton kind-indexed data family.

Universe Composition

data (f :.: g) a

Compose two Functors. Is the same as Compose and :.:, except with a singleton and meant to be used at the type level. Will be redundant if either of the above gets brought into the singletons library.

Note that because this is a higher-kinded data constructor, there is no SingKind instance; if you need fromSing and toSing, try going through Comp and getComp and SComp and sGetComp.

Note that Identity acts as an identity.

Since: 0.1.2.0

Constructors

Comp

Fields

getComp :: f (g a)

sGetComp :: Sing a -> Sing (GetComp a)

Singletonized witness for GetComp

Since: 0.1.2.0

type family GetComp c where ...

getComp lifted to the type level

Since: 0.1.2.0

Equations

GetComp (Comp a) = a

allComp :: (All f (All g p) @@ as) -> All (f :.: g) p @@ Comp as

Turn a composition of All into an All of a composition.

Since: 0.1.2.0

compAll :: (All (f :.: g) p @@ Comp as) -> All f (All g p) @@ as

Turn an All of a composition into a composition of All.

Since: 0.1.2.0

anyComp :: (Any f (Any g p) @@ as) -> Any (f :.: g) p @@ Comp as

Turn a composition of Any into an Any of a composition.

Since: 0.1.2.0

compAny :: (Any (f :.: g) p @@ Comp as) -> Any f (Any g p) @@ as

Turn an Any of a composition into a composition of Any.

Since: 0.1.2.0

Universe Disjunction

data (f :+: g) a

Disjoint union of two Functors. Is the same as Sum and :+:, except with a singleton and meant to be used at the type level. Will be redundant if either of the above gets brought into the singletons library.

Note that because this is a higher-kinded data constructor, there is no SingKind instance; if you need fromSing and toSing, consider manually pattern matching.

Note that Proxy acts as an identity.

Since: 0.1.3.0

Constructors

InL (f a)

InR (g a)

anySumL :: (Any f p @@ as) -> Any (f :+: g) p @@ InL as

Turn an Any of f into an Any of f :+: g.

anySumR :: (Any g p @@ bs) -> Any (f :+: g) p @@ InR bs

Turn an Any of g into an Any of f :+: g.

sumLAny :: (Any (f :+: g) p @@ InL as) -> Any f p @@ as

Turn an Any of f :+: g into an Any of f.

sumRAny :: (Any (f :+: g) p @@ InR bs) -> Any g p @@ bs

Turn an Any of f :+: g into an Any of g.

allSumL :: (All f p @@ as) -> All (f :+: g) p @@ InL as

Turn an All of f into an All of f :+: g.

allSumR :: (All g p @@ bs) -> All (f :+: g) p @@ InR bs

Turn an All of g into an All of f :+: g.

sumLAll :: (All (f :+: g) p @@ InL as) -> All f p @@ as

Turn an All of f :+: g into an All of f.

sumRAll :: (All (f :+: g) p @@ InR bs) -> All g p @@ bs

Turn an All of f :+: g into an All of g.

Defunctionalization symbols

data ElemSym0 (f :: Type -> Type) :: f k ~> (k ~> Type)

data ElemSym1 (f :: Type -> Type) :: f k -> k ~> Type

type ElemSym2 (f :: Type -> Type) (as :: f k) (a :: k) = Elem f as a

data GetCompSym0 :: (f :.: g) k ~> f (g k)

type GetCompSym1 a = GetComp a